Still working on 5.0, template system finished
Update gitignore

Remove test files

+++ TimoCloud-Universal/src/main/java/cloud/timo/TimoCloud/base/managers/BaseResourceManager.java
@@ -13,8 +13,8 @@ public class BaseResourceManager {
         this.operatingSystemMXBean = (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
     }
 
-    public int getFreeMemory() {
-        return (int) getOperatingSystemMXBean().getFreePhysicalMemorySize() / (1024*1024); // Convert to megabytes
+    public long getFreeMemory() {
+        return getOperatingSystemMXBean().getFreePhysicalMemorySize() / (1024*1024); // Convert to megabytes
     }
 
     public double getCpuUsage() {
